requirements.txt
asgiref==3.9.1
Django==5.2.4
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-tailwind==4.2.0
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
